fix process signaling
In order to get Docker to notify the service properly that it needs to quit,
ruby must be the main running process (or we should pass the signal somehow from
our start script).

`exec` is a command that replace the current process by another one.

NOTE: this only affect the jruby build
NOTE: pinned jruby to jruby 9.2 as that seems to be last compatible version with
current version of gems